Add files via upload
A web calculator is a digital tool available on a website that performs mathematical calculations. It allows users to input numerical data and execute various mathematical operations, such as addition, subtraction, multiplication, and division, as well as more advanced functions like logarithms, trigonometric functions, and exponentials. The output of the calculations is displayed on the screen, and users can use the calculator as many times as needed for various tasks. Web calculators can be used in a wide range of fields, including finance, engineering, science, and education.